Fix tuplePatternStart rule for [*foo] patterns
A tuple pattern may contain only a variadic variable (specified in
298aced). However, the `tuplePatternStart` lookahead rule previously
didn't account for this. The result of this was that the `pattern` rule
was unable to parse these patterns, attempting instead an interpretation
as `variablePattern`.

This can be seen, for instance, in this code:

    value v = let ([*e] = [0]) 0;

Without this fix, a syntax error occurs ("extraneous token 'e' expecting
closing bracket ']'", interpreting the `[` as the part of a tuple type
as in `[Integer] e = [0]`).

Destructuring statements with such patterns already work without this
fix because they use the `tupleOrEntryPattern` rule, where the last
alternative (chosen if lookahead for `tuplePatternStart` doesn't match)
is `tuplePattern`.